Dev/renzo

+++ app/Controller/ConvertController.php
@@ -8,16 +8,20 @@ use Flight;
 class ConvertController{
     public function convert(){
         $upload = Flight::request()->getUploadedFiles()['documentFile'];
+         if($upload->getClientMediaType() != "application/pdf"){
+            Flight::redirect('/?error-file');
+            die;
+        } 
         if ($upload->getError() === UPLOAD_ERR_OK) {
             $path = './files/'.$upload->getClientFilename();
             $upload->moveTo($path);
-            (new ConvertService())->transfor($path);
+            $textXML = (new ConvertService())->transfor($path);
             if (file_exists($path)) {
                 unlink($path);
             }
-            Flight::redirect('/?success');
+            Flight::redirect('/?success=1&total='.urlencode((string)$textXML));
         } else {
-            Flight::redirect('/');
+            Flight::redirect('/?error');
         }
     }
     public function download(){
@@ -33,7 +37,7 @@ class ConvertController{
             readfile($file);
             exit;
         } else {
-            Flight::redirect('/');
+            Flight::redirect('/?not-file');
         }
     }
 }
